Фикс серьезных багов покупки(возможно, проблемы были на стороне хостинга)

+++ modules/crypto_operations.py
@@ -68,19 +68,19 @@ class CryptoOperating:
         else:
             return response['balance'] / 1000000000000000000
 
-    def send_bitcoins(self, private_key: str, to_public_address: str, amount: int):
+    def send_bitcoins(self, private_key: str, to_public_address: str, amount: float):
         amount_to_satoshi = int(amount * 100000000)
         tx = simple_spend(private_key, to_public_address, amount_to_satoshi, coin_symbol='btc',
                           api_key=self.token, privkey_is_compressed=False)
         return tx
 
-    def send_ltc(self, private_key: str, to_public_address: str, amount: int):
+    def send_ltc(self, private_key: str, to_public_address: str, amount: float):
         amount_to_satoshi = int(amount * 100000000)
         tx = simple_spend(private_key, to_public_address, amount_to_satoshi, coin_symbol='ltc',
                           api_key=self.token, privkey_is_compressed=False)
         return tx
 
-    def send_doges(self, private_key: str, to_public_address: str, amount: int):
+    def send_doges(self, private_key: str, to_public_address: str, amount: float):
         amount_to_satoshi = int(amount * 100000000)
         tx = simple_spend(private_key, to_public_address, amount_to_satoshi, coin_symbol='doge',
                           api_key=self.token,
